+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import Header from "./Components/Header/Header";
 import NavBar from "./Components/Navbar/NavBar";
-import { Index } from "./Components/Content/Index";
+import { Index } from "./Pages/Index";
 import { SearchBar } from "./Components/Searchbar/SearchBar";
 import Footer from "./Components/Footer/Footer";
 import { Routes, Route } from "react-router-dom";
@@ -9,45 +9,35 @@ import Welcome from "./Pages/Welcome";
 import { About } from "./Pages/About";
 import MapWrapper from "./Components/map/MapWrapper";
 import View from "./Components/View/View";
-
+import Sidebar from "./Components/Sidebar/Sidebar";
 
 function App() {
   return (
     <>
       <header>
         <Header />
-      <NavBar />
-      <SearchBar />
-      <Index />
-      
+        <NavBar />
+        <SearchBar />
+        <Index />
       </header>
-      
-      
 
       <Routes>
-            <Route 
-            path='/' 
-            element={<Welcome/>} 
-            />
-            <Route 
-            path='/parks' 
-            element={<Index/>} 
-            />
-            <Route 
-            path='/about' 
-            element={<About/>} 
-            />
-            <Route
-            path='/map'
-            element={<MapWrapper/>}
-            />
-            <Route path="/:id/view"
-                    element={<View/>}
-                    />
-        </Routes>
+        <Route path="/" element={<Welcome />} />
+        <Route
+          path="/parks"
+          element={
+            <>
+              <Index />
+              <Sidebar />
+            </>
+          }
+        />
+        <Route path="/about" element={<About />} />
+        <Route path="/map" element={<MapWrapper />} />
+        <Route path="/:id/view" element={<View />} />
+      </Routes>
 
-       
-        <Footer/>
+      <Footer />
     </>
   );
 }
